+++ modules/tinymce/src/themes/silver/main/ts/ui/dialog/BodyPanel.ts
@@ -1,4 +1,4 @@
-import { Behaviour, Form as AlloyForm, Keying, Memento, SimpleSpec } from '@ephox/alloy';
+import { AddEventsBehaviour, AlloyEvents, Behaviour, Form as AlloyForm, Keying, Memento, NativeEvents, SimpleSpec } from '@ephox/alloy';
 import { Dialog } from '@ephox/bridge';
 import { Arr, Fun, Optional } from '@ephox/katamari';
 
@@ -8,6 +8,7 @@ import { RepresentingConfigs } from '../alien/RepresentingConfigs';
 import * as FormValues from '../general/FormValues';
 import * as NavigableObject from '../general/NavigableObject';
 import { interpretInForm } from '../general/UiFactory';
+import { dialogFocusShiftedChannel } from '../window/DialogChannels';
 
 export type BodyPanelSpec = Omit<Dialog.Panel, 'type'>;
 
@@ -55,7 +56,15 @@ const renderBodyPanel = (spec: BodyPanelSpec, dialogData: Dialog.DialogData, bac
           },
           Fun.identity
         )
-      })
+      }),
+      AddEventsBehaviour.config('dialog-body-panel', [
+        // TINY-10101: This is to cater the case where clicks are made into the dialog instead using keyboard navigaton, as FocusShifted would not be triggered in that case.
+        AlloyEvents.run(NativeEvents.focusin(), (comp, se) => {
+          comp.getSystem().broadcastOn([ dialogFocusShiftedChannel ], {
+            newFocus: Optional.some(se.event.target)
+          });
+        }),
+      ])
     ])
   };
 };