+++ modules/tinymce/src/themes/silver/main/ts/ui/dialog/IFrame.ts
@@ -1,17 +1,22 @@
-import { AlloyComponent, Behaviour, Focusing, FormField, SketchSpec, Tabstopping } from '@ephox/alloy';
+import { AlloyComponent, Behaviour, Focusing, FormField, Receiving, SketchSpec, Tabstopping } from '@ephox/alloy';
 import { Dialog } from '@ephox/bridge';
 import { Cell, Fun, Optional, Throttler, Type } from '@ephox/katamari';
 import { PlatformDetection } from '@ephox/sand';
-import { Attribute, SugarElement } from '@ephox/sugar';
+import { Attribute, Class, Compare, SugarElement, Traverse } from '@ephox/sugar';
 
 import { UiFactoryBackstageProviders } from '../../backstage/Backstage';
 import { renderFormFieldWith, renderLabel } from '../alien/FieldLabeller';
 import { RepresentingConfigs } from '../alien/RepresentingConfigs';
 import * as NavigableObject from '../general/NavigableObject';
+import { dialogFocusShiftedChannel } from '../window/DialogChannels';
 
 interface IFrameSourcing {
   readonly getValue: (frame: AlloyComponent) => string;
   readonly setValue: (frame: AlloyComponent, value: string) => void;
+}
+
+export interface DialogFocusShiftedEvent extends CustomEvent {
+  readonly newFocus: Optional<SugarElement<HTMLElement>>;
 }
 
 type IframeSpec = Omit<Dialog.Iframe, 'type'>;
@@ -111,13 +116,13 @@ const getDynamicSource = (initialData: Optional<string>, stream: boolean): IFram
 
 const renderIFrame = (spec: IframeSpec, providersBackstage: UiFactoryBackstageProviders, initialData: Optional<string>): SketchSpec => {
   const baseClass = 'tox-dialog__iframe';
-  const borderedClass = spec.border ? [ `${baseClass}--bordered` ] : [];
   const opaqueClass = spec.transparent ? [] : [ `${baseClass}--opaque` ];
+  const containerBorderedClass = spec.border ? [ `tox-navobj-bordered` ] : [];
 
   const attributes = {
     ...spec.label.map<{ title?: string }>((title) => ({ title })).getOr({}),
     ...initialData.map((html) => ({ srcdoc: html })).getOr({}),
-    ...spec.sandboxed ? { sandbox: 'allow-scripts allow-same-origin' } : { },
+    ...spec.sandboxed ? { sandbox: 'allow-scripts allow-same-origin' } : { }
   };
 
   const sourcing = getDynamicSource(initialData, spec.streamContent);
@@ -125,6 +130,7 @@ const renderIFrame = (spec: IframeSpec, providersBackstage: UiFactoryBackstagePr
   const pLabel = spec.label.map((label) => renderLabel(label, providersBackstage));
 
   const factory = (newSpec: { uid: string }) => NavigableObject.craft(
+    Optional.from(containerBorderedClass),
     {
       // We need to use the part uid or the label and field won't be linked with ARIA
       uid: newSpec.uid,
@@ -133,14 +139,27 @@ const renderIFrame = (spec: IframeSpec, providersBackstage: UiFactoryBackstagePr
         attributes,
         classes: [
           baseClass,
-          ...borderedClass,
           ...opaqueClass
         ]
       },
       behaviours: Behaviour.derive([
         Tabstopping.config({ }),
         Focusing.config({ }),
-        RepresentingConfigs.withComp(initialData, sourcing.getValue, sourcing.setValue)
+        RepresentingConfigs.withComp(initialData, sourcing.getValue, sourcing.setValue),
+        Receiving.config({
+          channels: {
+            [dialogFocusShiftedChannel]: {
+              onReceive: (comp, message: DialogFocusShiftedEvent) => {
+                message.newFocus.each((newFocus) => {
+                  Traverse.parentElement(comp.element).each((parent) => {
+                    const f = Compare.eq(comp.element, newFocus) ? Class.add : Class.remove;
+                    f(parent, 'tox-navobj-bordered-focus');
+                  });
+                });
+              }
+            }
+          }
+        })
       ])
     }
   );